+++ zapdos/pokemongo/admin.py
@@ -5,8 +5,6 @@ from django.utils.translation import pgettext_lazy as pgettext
 from pokemongo.constants import BATTLE_HUB_STATS, STANDARD_MEDALS, UPDATE_FIELDS_TYPES
 from pokemongo.models import (
     BattleHubPost,
-    Community,
-    Faction,
     FactionAlliance,
     GymBadgePost,
     MedalProgressPost,
@@ -15,13 +13,6 @@ from pokemongo.models import (
 admin.site.register(FactionAlliance)
 admin.site.register(BattleHubPost)
 admin.site.register(GymBadgePost)
-
-
-@admin.register(Community)
-class CommunityAdmin(admin.ModelAdmin):
-
-    search_fields = ("name", "short_description", "handle")
-    autocomplete_fields = ["memberships"]
 
 
 @admin.register(MedalProgressPost)
